test(cf-t8k1.fu1): split lazy-init tests into 2 files per noDoMockInTestBody invariant
melania CR escalation on PR #1364: the legacy-tag escape hatch isn't
appropriate for NEW code. The convention (CF-fgsw, tests/
noDoMockInTestBody.test.js) requires top-level vi.mock() for new tests
+ legacy tags only for pre-existing violations being grandfathered.

## Refactor

The 2 describe blocks in upsShippingLazyInit.test.js needed DIFFERENT
mock shapes for public/sharedTokens (empty for the import-resilience
contract; populated for the getter-shape contract). vi.doMock() in
beforeEach was the wrong tool — vi.doMock doesn't hoist + doesn't
invalidate already-resolved static imports.

Solution: split into 2 files, each with module-level vi.mock declaring
its own mock shape once.

- `tests/upsShippingLazyInit.test.js` — EMPTY mock; pins the
  import-resilience contract (module loads + getter throws-at-call-time)
- `tests/upsShippingLazyInitPopulated.test.js` — POPULATED mock; pins
  the getter-shape + memoization invariants

## Verification

- 7 tests across both files, all pass
- noDoMockInTestBody.test.js (CF-fgsw invariant): now compliant
- 76 broader shipping tests still green (verified locally before push)
- No production code change vs PR #1364's earlier state

Test-only refactor; prior approvals (godfrey, radahn, rennala) stand —
the lazy-init source change is unchanged.
